+++ src/main/java/dansplugins/dansessentials/commands/BackCommand.java
@@ -29,8 +29,14 @@ public class BackCommand extends AbstractPluginCommand {
 
         Player player = (Player) commandSender;
 
+        // Check if we have a previous location stored
+        if (ephemeralData.getLastLogins().get(player) == null) {
+            player.sendMessage(ChatColor.RED + "You don't have a previous location to return to!");
+            return false;
+        }
+
         player.teleport(ephemeralData.getLastLogins().get(player));
-        player.sendMessage(ChatColor.AQUA + "Teleported to your last location!");
+        player.sendMessage(ChatColor.AQUA + "Teleported to your previous location!");
         return true;
     }